#300: Refactoring to support more focus in testing
