+++ CourageScores/ClientApp/src/components/division_fixtures/scores/MergeMatch.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import {useApp} from "../../../AppContainer";
+import {matchEquals} from "./MatchComparer";
 
 export function MergeMatch({ readOnly, matches, matchIndex, homeSubmission, awaySubmission, setFixtureData, fixtureData }) {
     const { onError } = useApp();
@@ -20,46 +21,6 @@ export function MergeMatch({ readOnly, matches, matchIndex, homeSubmission, away
         } catch (e) {
             onError(e);
         }
-    }
-
-    function matchEquals(x, y) {
-        if (!x && !y) {
-            return true;
-        }
-
-        if (!y || !x) {
-            return false;
-        }
-
-        return x.homeScore === y.homeScore
-            && x.awayScore === y.awayScore
-            && playersEqual(x.homePlayers, y.homePlayers)
-            && playersEqual(x.awayPlayers, y.awayPlayers);
-    }
-
-    function playersEqual(xPlayers, yPlayers) {
-        if (!xPlayers && !yPlayers) {
-            return true;
-        }
-
-        if (!xPlayers || !yPlayers) {
-            return false;
-        }
-
-        if (xPlayers.length !== yPlayers.length) {
-            return false;
-        }
-
-        for (let index = 0; index < xPlayers.length; index++) {
-            const xPlayer = xPlayers[index];
-            const yPlayer = yPlayers[index];
-
-            if (xPlayer.id !== yPlayer.id) {
-                return false;
-            }
-        }
-
-        return true;
     }
 
     function combinePlayers(homePlayers, awayPlayers) {
